♻️[Refactor] Extension 분리 및 cell 폴더 네이밍 변경
